Changed the HTTP server from http.server to aiohttp
This server is compatible with the socket.io server I am planning to
use (python-socketio). It is also a lot easier to work with than
http.server

The HTTP_Presentation class has been changed accordingly, along with
the initialisation code in the "serve" module.

The commandline interface is still fully compatible with previous
versions of the program.

The hashbang has also been changed to accomodate the use of
virtualenv, so that the kernel select the python3 interpreter which
comes with virtualenv, instead of the "global" one.