Added a 'local' provider and the ability to override presentation providers
The local provider is meant to serve the Reveal resources from the host
the presentation(s) are served from. One can provide the program with
a repository, or it can search for one itself (using a few builtin paths
to look for it. Another possibility is that another webserver serves the
files for WaterSlide, under /reveal.js/ at the same hostname.

The override option can be used to override the configured resource provider.
This might be used in a situation where the presentation wanted the resources
to be fetched from, e.g.: cloudflare, but the internet is down. But when a
local repository is available, one can override the configuration and let the
links point to the local repository.

Any combination of the above two features can be used.

The documentation has been rewritten where necessary to reflect these
new features and the possibilities they provide.

Another change this commit brings with it is that we fully use the
virtualenv/pip/requirements.txt suite of tools.