Added a basic multiplexing implementation.
For Socket.io, we use the socketio-python library. The implementation
of the server side is largely mirrored upon the node.js implementation.

Multiplexing is currently enabled on the server by passing the -m flag
to the serve subcommand. In the browser, we query for the master status
(for example; /?master). It isn't really secure yet (anyone can become
the master now), but when that is solved, it should be secure.

The multiplexing currently suffers from feedback loops when more than
one master is connected, and one master changes state at a high
frequency. This should be fixed in the future.