Made sure the presenation won't spaz out by preventing feedback loops.
The system now follows the "Thou shall not rebroadcast events to it's
origin, and prevent rebroadcasts from received events" commandment.

This ensures that feedback loops can no longer build up. The only thing
that happens now is that if you flip between slides at a sufficiently high
frequency, the other presentations will start to lag, but if you stop they
will eventually work of their queues and come to a halt at the same slide
the master is currently at. This should not be a problem in normal use.