Added the infrastructure to create a package for waterslide
To accomodate packaging, the entry point of the development mode has
shifted to the wslide script in the top level directory. That script
does not get added to the package, since pip creates a script for us.

Also, to get waterslide recognised as a module, I added a __init__.py
file. Also, the source directory has been renamed from src to waterslide.