Simplified the routing scheme by eliminating redundant routes and logic
When more than one presentation is being served, a redirect route is
necessary for when the file instead of the index is being requested
('/foo' vs '/foo/'). This was previously handled in the presentation
class, but it is now handled by a variable route. The now redundant
logic in the presentation class has been removed too

Other routes have also been replace by variable routes. The
'/', and '/{file}' routes have been replaced with the '/{tail:.*}'
route, which works in both cases.