Added a middleware class to keep HTTP_Presentation framework agnostic.
This middleware class converts between http request and response
representations. It is also supposed to rewrite all the paths which
get passed to presentation objects to be relative to that presentations
directory.

consider /foo/bar.js within the foo presentation. The url which is then
passed to the foo object should be 'bar.js'.