send_direct() now reads as binary.
Before it decoded files first to utf-8, which wasn't a problem for
text files, but it was a problem for images.

Luckily, we can just put utf8 text in the body of a response,
without having to decode it to bytes first. We can also then just
pass the image bytes directly into it. Now the send direct method can
just send binary.