Made caching work AGAIN by flooring the timestamp to seconds.
The filesystem returns a timestamp with microsecond precision, while
the browser gives us one with only seconds of precision. When we
attempt to compare them, we fail the comparison because the one has
microseconds, and the other has not.

When we floor the microseconds, we can compare them sucessfully. We
could also round them, but in the end it has the same effect. This is
because when we send the timestamp for the browser to remember (whether
it was floored or rounded), and then redo the comparison without the
file being changed, we successfully do the comparison, because neither
timestamp has changed.