Added a feature to disable static routes, and a configuration dump subcommand.
Disabling static routes will cause WaterSlide to return 403's when a user
agent tries to access a location which is stored on the filesystem and does
not require processing. This is then the responsibility of another webserver
(which can also be used to front-end WaterSlide, or encrypt the traffic,
or whatever).