Written a framework to manage a dynamic document root full of presentations
On a server it is of course advantageous not having to restart the service
whenever a key part of your content changed. This commit solved that problem

At the centerpiece is the Manager class, and its module. It manages
requests for presentations and other types of dynamic content under
its document root. To do this efficiently, a few other systems have been
developed to streamline this new system.

CACHING
------
First, the caching mechanism. It is applied as a decorator to functions
which output need to be cached. It is a rather general caching mechanism,
which could theoretically be used to cache almost anything. It also provides
a way to check (on each cach retrieval) if the cached resource is stale or not.
The manager class provides as of yet no way to set the caching depth of these
decorators, but that is planned

The way caching is used in the manager class is to use functions which only
generate and initialise objects which store the content itself. This way the
initialisation happens infrequently. These initialisation functions are called
through their decorators by the entry functions. These entry functions call the
handling functions on the objects returned. If a resource does not exist, the
initialisation function generates a "notfound" object, which can be transparently
handled by the entry functions. It is also expected that each object keeps an eye
on it's on-disk counterpart (or source file, if you will) to check that we don't
keep serving resources which have already disappeared from the backing store.

HTTP related changes
-------
Since the presentation module no longer has a monopoly on the HTTP protocol, most
of the functionality which was not strongly bound to the presentation module has
been moved to the new httputils module.

New functionality, such as the aio-translation decorator, is also located there.
This decorator translates the requests and responses from the server's
representation to our own and back again. It is configurable by providing a rewrite
and a logging function

Presentation changes
-------
The presentation class's initialisation routines now raise an ImportError exception
when something goes wrong while importing, instead of tediously returning False
through three layers of function.

Furthermore, the url is no longer based upon the title. This was incompatible with
the way the Manager class serves requests, and has therefore been removed to prevent
path conflicts between the exact same presentations whilst they are served in either
the 'serve' subcommand, or with the 'manage' subcommand.