Moved argument parsing to their associated configuration, added multiplexing to manager
Since some arguments are shared between the serve and the manage subcommand, I
have moved the parsing of those arguments into their associated configuration
objects. The helptexts have also been moved into those objects

There was also a bug in the part of the multiplexing logic which determined
whether or not the current request may be a master. It didn't remove the
secret.

Furthermore, to ensure that all subcommands which serve presentations set up
a default enviroment, the startup_default() function in the httputils module
can be called. It sets up the static routes, and, when necessary, starts up the
multiplexing subsystem. The static route initialisation now also lives in httputils.