Added autoslaving control to multiplexing.
Previously all clients which weren't masters were automatically slaved
to the master presentation. This functionality can now be controlled by
passing either --autoslave or --no-autoslave to the program, or relying
on the default settings (seve will default to autoslave, manage will not)

This is primarily useful for server deployments, where it isn't intended
that unsuspecting viewers get slaved to a master which is using the same
presentation elsewhere, and then get dragged with said presentation.