Changed multiplexing logic to be on demand, and with sha512 insteadof bcrypt
The multiplexing dictionary is now calculated upon each request, instead
of being prepared upon initialisation. This is because we can then run
multiple multiplexing presentations next to each other without interfering
using "sessions"

Sessions, in this case, is to send a session identifier to the server. That
identifier can be send using a url parameter after "master", or "slave"
This identifier then get's concatenated to the standard randomness, after
which it is hashed.

The hashing algorithm has also changed from bcrypt to sha512. It is faster,
because we don't have to perform a _very_ secure (and therefore slow) password
hash. Since we now use a much faster algorithm (it doesn't have to be blazingly
secure), the performance has increased noticeably. Where before it took a while
for the other clients to catch up (as in, starting to move after the slide on the
master was already done), it now moves instantly.

I also removed all the requirements which were no longer necessary due to the
removal of bcrypt from setup.py and the requirements file.