Added master authorisation through BasicAuth, added some extra comments
The credentials are stored in the conf.yaml file of the presentation,
under multiplex > auth > (uname, passwd), both in plain text ^1. The goal
of this is to prevent some aspiring dictators from noticing the url in the
url bar and trying out if they can hijack the presentation.

This commit also adds a commandline flag to enable tracing of multiplex
frames.

^1:	Yes I know it isn't secure, but that isn't the point. It isn't if
	there is anything sensitive stored behind that password, and I
	wanted to be able to easily change the password without having
	to involve another program.