Added favicon handling
One can now set the favicon to be used for the presentation using
the 'favicon' configuration parameter in a presentation. If it is not
set, it will link to the waterslide logo (which is shipped with each
release). The configuration documentation has been changed accordingly.

The logo itself is developed in the logo directory, where it is transpiled
to .png and a plain .svg with the Makefile in that directory, and then copied
to the web-resources directory by the global makefile.